order create

+++ jd-store/src/main/java/com/store/jdstore/entity/HanmaAddressEntity.java
@@ -18,7 +18,7 @@ public class HanmaAddressEntity {
   private String city;
   private String area;
   private String house;
-  private Long isDefault;
+  private Integer isDefault;
   private String createTime;
   private String updateTime;
