+++ jd-store/src/main/java/com/store/jdstore/entity/HanmaShopEntity.java
@@ -3,6 +3,8 @@ package com.store.jdstore.entity;
 import lombok.Data;
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
+
 
 @Data
 @Entity
@@ -16,11 +18,11 @@ public class HanmaShopEntity {
     private String name;
     private Long sales;
     private Long expressLimit;
-    private Double expressPrice;
+    private BigDecimal expressPrice;
     private String desc;
     private String imgUrl;
     private Long state;
-    private Long status;
+    private Integer status;
 
     private String createTime;
