+++ jd-store/src/main/java/com/store/jdstore/model/ProductModel.java
@@ -2,6 +2,8 @@ package com.store.jdstore.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 
 @Data
 public class ProductModel {
@@ -9,12 +11,12 @@ public class ProductModel {
   private Long id;
   private String name;
   private Long sales;
-  private Double price;
-  private Double oldPrice;
+  private BigDecimal price;
+  private BigDecimal oldPrice;
   private Long productStock;
   private String imgUrl;
   private String type;
   private Long shopId;
-  private Long status;
+  private Integer status;
 
 }
